+++ NEWS.md
@@ -8,6 +8,11 @@ News
 
 * Added argument `verbose` in `nestcv.train()` and `nestcv.glmnet()` to show 
 progress.
+* Added argument `multicore_fork` in `nestcv.train()` to allow choice of 
+parallelisation between forked multicore processing using `mclapply` or 
+non-forked using `parLapply`. This can help prevent errors with some 
+multithreaded caret models e.g. `model = "xgbTree"` when `tuneLength` is set to 
+a high number.
 * In `one_hot()` changed `all_levels` argument default to `FALSE` to be 
 compatible with regression models by default.